Upload first working version
The files for the first Phase 4 program, including the main targeting program: `control.lua`, a helper program to reset the computer's redstone outputs: `reset.lua`, and the schematic for my build so you can place it down and try it out yourself, which is the `control_schematic.nbt` file. Schematics can be brought into your world using the schematic table and empty schematic from the Create mod.
I deviated slightly from the original development procedure because the math turned out to be very hard to solve algebraically. Instead of solving for the initial angle, I simulated firing the shots using the same code the mod itself uses, and altered the chosen trajectory until the shot landed within a specified distance of the target. I also did not make a separate range calculator, but I may add that to this repository later.