Initial UI program upload
Initial version of the UI program involves calling "ui.lua" on the command line while a 4x3 (4wide, 3 tall) monitor is connect to the computer via a wired modem. This also assumes you are using the included schematic for the cannon control system, and "aim.lua" is also present in the same directory as "ui.lua". The only button on the main menu that works is "AIM", and pressing that brings you to the aiming menu. From there, you select the coordinate boxes and use the numpad and negate, backspace, and clear to set the x, y, and z values. You can choose to use relative or exact coordinates, and the default is exact (both start as grayed-out, but if none are chosen then exact is assumed). Finally, assemble the cannon manually (with the linked controller), and then press "Aim at Target".